// src/Models/PreContentieuxType.ts
export interface PreContentieuxType {
  id: number;
  preContentieuxTypeCategorieRef: number;
  preContentieuxTypeLibelle: string;
  preContentieuxTypeLibelleArabe: string;
  preContentieuxTypeDescription: string;
  preContentieuxTypeDescriptionArabe: string;
  preContentieuxTypeDelai: number;
  preContentieuxTypeEtat: number;
}